# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Uni Ride Sharing API",
    "description": "API documentation for the university ride sharing application",
    "version": "1.0.0"
  },
  "host": "localhost:3002",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "sessionAuth": {
      "type": "apiKey",
      "name": "Cookie",
      "in": "header",
      "description": "Session cookie for authentication"
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "User login",
        "description": "Authenticates a user and creates a session",
        "tags": ["Authentication"],
        "parameters": [
          {
            "name": "credentials",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["email", "password", "userType"],
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email",
                  "example": "john@example.com"
                },
                "password": {
                  "type": "string",
                  "format": "password",
                  "example": "password123"
                },
                "userType": {
                  "type": "string",
                  "enum": ["student", "driver", "admin"],
                  "example": "student"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
...
    "/messages": {
      "get": {
        "summary": "Get message threads",
        "description": "Gets all message threads for the current user",
        "tags": ["Messages"],
        "security": [{"sessionAuth": []}],
        "responses": {
          "200": {
            "description": "List of message threads"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      
      "post": {
        "summary": "Send a message",
        "description": "Sends a new message to another user",
        "tags": ["Messages"],
        "security": [{"sessionAuth": []}],
        "parameters": [
          {
            "name": "messageData",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["recipient", "subject", "content"],
              "properties": {
                "recipient": {
                  "type": "string",
                  "example": "passenger:123"
                },
                "subject": {
                  "type": "string",
                  "example": "Regarding ride on Friday"
                },
                "content": {
                  "type": "string",
                  "example": "Hi, I'd like to confirm our ride for Friday."
                },
                "rideId": {
                  "type": "integer",
                  "example": 456
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Message sent successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
} 